🚒🏬 ↝ [SSG-287]: Reorganising some of the satellite/project classification components
