+++ app/next/[id]/page.tsx
@@ -1,4 +1,4 @@
-import ClientClassificationPage from "@/src/components/research/projects/(classifications)/NextScene";
+import ClientClassificationPage from "@/src/components/projects/(classifications)/NextScene";
 
 type Props = {
   params: { id: string };
