+++ app/structures/balloon/[project]/[id]/[mission]/page.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react"
 import { useParams, useRouter } from "next/navigation"
 import { useSessionContext } from "@supabase/auth-helpers-react"
 
-import { StarterLidar } from "@/src/components/research/projects/Lidar/Clouds"
+import { StarterLidar } from "@/src/components/projects/Lidar/Clouds"
 import GameNavbar from "@/src/components/layout/Tes"
 
 export default function BalloonClassifyPage() {
