+++ src/components/deployment/missions/structures/Astronomers/SatellitePhotos/AI4M/AIForMars.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { useSupabaseClient, useSession } from "@supabase/auth-helpers-react";
 import MissionShell from "../../../BasePlate";
 import { CloudCogIcon, FolderCog, HelpCircle, PaintBucket, Vote } from "lucide-react";
-import { AiForMarsProjectWithID, StarterAiForMars } from "@/src/components/research/projects/Auto/AI4Mars";
+import { AiForMarsProjectWithID, StarterAiForMars } from "@/src/components/projects/Auto/AI4Mars";
 import VoteAI4MClassifications from "./AI4MVote";
 
 export interface Mission {
