+++ src/components/projects/(classifications)/Annotating/DrawingCanvas.tsx
