+++ src/components/projects/(classifications)/Annotating/DrawingUtils.tsx
