+++ src/components/projects/(classifications)/NextScene.tsx
