+++ src/components/projects/Lidar/Clouds.tsx
@@ -3,7 +3,6 @@
 import React, { useEffect, useState } from "react"; 
 import { useSupabaseClient, useSession } from "@supabase/auth-helpers-react";
 import { useActivePlanet } from "@/src/core/context/ActivePlanet";
-import ClassificationForm from "@/src/components/research/projects/(classifications)/PostForm";
 import { CloudspottingOnMarsTutorial } from "./cloudspottingOnMars"; 
 import ImageAnnotator from "../(classifications)/Annotating/Annotator";
