+++ src/components/projects/Lidar/EarthCloudsRead.tsx
@@ -3,7 +3,6 @@
 import React, { useEffect, useState } from "react";
 import { useSupabaseClient, useSession } from "@supabase/auth-helpers-react";
 import { useActivePlanet } from "@/src/core/context/ActivePlanet";
-import ClassificationForm from "@/src/components/research/projects/(classifications)/PostForm";
 
 export function EarthCloudRead() {
     const supabase = useSupabaseClient();
