+++ src/components/projects/Telescopes/AreWeAlone.tsx
