+++ src/components/projects/Telescopes/DailyMinorPlanet.tsx
