+++ src/components/projects/Telescopes/Sunspots.tsx
@@ -3,7 +3,7 @@
 import React, { useEffect, useState } from "react";
 import { useSupabaseClient, useSession } from "@supabase/auth-helpers-react";
 import { useActivePlanet } from "@/src/core/context/ActivePlanet";
-import ClassificationForm from "@/src/components/research/projects/(classifications)/PostForm";
+import ClassificationForm from "../(classifications)/PostForm";
 
 interface SelectedAnomalyProps {
     anomalyid: number;
