+++ src/components/projects/Telescopes/Sunspots/SunspotShell.tsx
