+++ src/components/projects/Telescopes/Sunspots/SunspotVote.tsx
